+++ NN_2L2H2O_poly2d_CPU/src/utility.h
@@ -49,15 +49,7 @@ using matrix_by_vector_t = std::vector<std::vector<T> >;
 //
 // Check if a string is a float number
 template <typename T>
-bool IsFloat( std::string myString ) {
-    std::istringstream iss(myString);
-    T f;
-    iss >> std::skipws >> f; // skipws ignores leading whitespace
-    // Check the entire string was consumed and if either failbit or badbit is set
-    return iss.eof() && !iss.fail(); 
-}
-
-
+bool IsFloat( std::string myString );
 
 
 //==============================================================================
@@ -65,69 +57,23 @@ bool IsFloat( std::string myString ) {
 // Memory management functions
 //
 template <typename T>
-void clearMemo(T** & data){
-     if (data != nullptr) {
-          if (data[0] != nullptr)  delete[] data[0];   
-          delete[] data;
-     };
-     data = nullptr;
-     return;
-}
+void clearMemo(T** & data);
+
+template <typename T>
+void clearMemo(std::vector<T**> & data);
 
 
 template <typename T>
-void clearMemo(std::vector<T**> & data){
-     for (auto it=data.rbegin(); it!=data.rend(); it++){
-          clearMemo<T>(*it);
-          data.pop_back();
-     }
-     return;
-}
-
-
-
-template <typename T>
-void clearMemo(std::map<std::string, T**> & data){
-     for (auto it=data.begin(); it!=data.end(); it++){
-          clearMemo<T>(it->second);
-          it->second = nullptr;
-     }
-     return;
-}
-
-
-
-
+void clearMemo(std::map<std::string, T**> & data);
 
 
 
 //==============================================================================
 //
 // Initialize a matrix in consecutive memory
+
 template <typename T>
-bool init_mtx_in_mem(T** & data, size_t rows, size_t cols){
-     try{          
-          //clearMemo<T>(data);
-          if( rows*cols >0) {          
-               T * p = new T[rows*cols]();
-               data = new T* [rows];
-               #ifdef _OPENMP
-               #pragma omp parallel for shared(data, p, rows, cols)
-               #endif 
-               for(int ii=0; ii<rows; ii++){                    
-                    data[ii]=p+ii*cols;     
-                    memset(data[ii], 0, sizeof(T)*cols);       
-               }          
-          }     
-          return true;
-     } catch (...){
-          clearMemo<T>(data);   
-          return false;
-     }
-};
-
-
-
+bool init_mtx_in_mem(T** & data, size_t rows, size_t cols);
 
 
 
@@ -135,115 +81,13 @@ bool init_mtx_in_mem(T** & data, size_t rows, size_t cols){
 //
 // Read in a 2D array from file and save to  **data / rows / cols
 template <typename T>
-int read2DArrayfile(T** & data, size_t& rows, size_t& cols, const char* file, int titleline=0){
-    try { 
-          
-          clearMemo<T>(data);
-          
-          std::ifstream ifs(file);
-          std::string line;
-          matrix_by_vector_t<T> mtx;
-          
-          for (int i=0; i < titleline; i++){          
-               getline(ifs,line);
-          }
-          std::vector<T> onelinedata;
-          while(getline(ifs, line)){          
-               char* p = &line[0u];
-               char* end;              
-               onelinedata.clear();                              
-               for( T d = strtod(p, &end); p != end; d = strtod(p, &end) ) {
-                    p = end;                    
-                    onelinedata.push_back(d);           
-               };               
-               if(onelinedata.size()>0) mtx.push_back(onelinedata);              
-          }          
-
-          rows=mtx.size();
-          if (rows > 0){
-               cols=mtx[0].size();
-               
-               init_mtx_in_mem<T>(data, rows, cols);  
-               
-               #ifdef _OPENMP
-               #pragma omp parallel for simd shared(data, mtx, rows)
-               #endif                                      
-               for(int ii=0; ii<rows; ii++){
-                    copy(mtx[ii].begin(), mtx[ii].end(), data[ii]);       
-               }          
-          } else {
-               std::cout << " No Data is read from file as 2D array" << std::endl;
-          }
-          mtx.clear();
-          return 0;                    
-    } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-        return 1;
-    }
-};
+int read2DArrayfile(T** & data, size_t& rows, size_t& cols, const char* file, int titleline=0);
 
 
 template <typename T>
-int read2DArray_with_max_thredhold(T** & data, size_t& rows, size_t& cols, const char* file, int titleline=0, int thredhold_col=0, T thredhold_max=std::numeric_limits<T>::max()){
-    try { 
-          
-          clearMemo<T>(data);
-          
-          std::ifstream ifs(file);
-          std::string line;
-          matrix_by_vector_t<T> mtx;
-          
-          for (int i=0; i < titleline; i++){          
-               getline(ifs,line);
-          }
-          std::vector<T> onelinedata;
-          while(getline(ifs, line)){          
-               char* p = &line[0u];
-               char* end;              
-               onelinedata.clear();                              
-			
-               for( T d = strtod(p, &end); p != end; d = strtod(p, &end) ) {
-                    p = end;                  
-                    onelinedata.push_back(d);           
-               };               
-               if (onelinedata.size()>0) {   
-                    int checkcol = onelinedata.size() ;                                                        
-                    if ( thredhold_col >=0) {           
-                         // when thredhold_index is non-negative, check the colum VS max
-                         checkcol = thredhold_col;
-                    } else {
-                         // when thredhold_index is negative, check the column from the end VS max
-                         checkcol += thredhold_col;                                        
-                    }                         
-                    if (onelinedata[checkcol] > thredhold_max) {
-                         continue;  // If the data exceeds thredhold, ignore this line.
-                    }
-                    mtx.push_back(onelinedata);                               
-               }                           
-          }          
+int read2DArray_with_max_thredhold(T** & data, size_t& rows, size_t& cols, const char* file, int titleline=0,
+    int thredhold_col=0, T thredhold_max=std::numeric_limits<T>::max());
 
-          rows=mtx.size();
-          if (rows > 0){
-               cols=mtx[0].size();
-               
-               init_mtx_in_mem<T>(data, rows, cols);  
-               
-               #ifdef _OPENMP
-               #pragma omp parallel for simd shared(data, mtx, rows)
-               #endif                                      
-               for(int ii=0; ii<rows; ii++){
-                    copy(mtx[ii].begin(), mtx[ii].end(), data[ii]);       
-               }          
-          } else {
-               std::cout << " No Data is read from file as 2D array" << std::endl;
-          }
-          mtx.clear();
-          return 0;                    
-    } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-        return 1;
-    }
-};
 
 
 
@@ -251,27 +95,7 @@ int read2DArray_with_max_thredhold(T** & data, size_t& rows, size_t& cols, const
 // 2D array transpose
 //
 template <typename T>
-void transpose_mtx(T** & datdst,  T** datrsc,  size_t nrow_rsc, size_t ncol_rsc)
-{
-     try{ 
-     
-          if ( datdst== nullptr) init_mtx_in_mem<T>(datdst, ncol_rsc, nrow_rsc);
-          
-          // best way to transpose a c++ matrix is copying by row    
-          #ifdef _OPENMP
-          #pragma omp parallel for simd shared(datrsc, datdst, nrow_rsc, ncol_rsc)
-          #endif      
-          for(int irow = 0; irow< nrow_rsc; irow++){          
-               for(int icol=0; icol < ncol_rsc; icol++){
-                    datdst[icol][irow] = datrsc[irow][icol];      // not a smart way.
-               }
-          }          
-          
-     } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-     }
-};
-
+void transpose_mtx(T** & datdst,  T** datrsc,  size_t nrow_rsc, size_t ncol_rsc);
 
 
 
@@ -286,7 +110,6 @@ void transpose_mtx<float>(float** & datdst,  float** datrsc, size_t nrow_rsc, si
 #endif
 
 
-
 //===============================================================================                                     
 //
 // Matrix normalization utility functions
@@ -294,24 +117,7 @@ size_t get_count_by_percent(size_t src_count, double percentage);
                                      
                                      
 template<typename T>
-void get_max_each_row(T*& rst,  T* src,  size_t src_rows, size_t src_cols, long int col_start=0, long int col_end=-1){ 
-          if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive        
-          if(rst == nullptr) rst = new T[src_rows]();               
-          #ifdef _OPENMP
-          #pragma omp parallel for simd shared(src, rst, src_rows, src_cols, col_start, col_end)
-          #endif   
-		int idxlast = 0;
-          for(size_t ii=0 ; ii< src_rows ; ii++ ){
-               for(size_t jj=col_start; jj<= col_end ; jj++){
-                    if ( rst[ii] < fabs(src[ii*src_cols + jj]) ) {
-                         rst[ii] = fabs(src[ii*src_cols + jj]);
-					if(ii==(src_rows-1)){
-						idxlast = jj;
-					}
-                    };
-               }
-          };  
-};
+void get_max_each_row(T*& rst,  T* src,  size_t src_rows, size_t src_cols, long int col_start=0, long int col_end=-1);
 
 
 #if defined (_USE_GSL) || defined (_USE_MKL)
@@ -325,19 +131,9 @@ void get_max_each_row<float>(float*& rst, float* src, size_t src_rows, size_t sr
 
 
 template<typename T>
-void norm_rows_in_mtx_by_col_vector(T* src_mtx, size_t src_rows, size_t src_cols, T* scale_vec, long int col_start=0, long int col_end=-1 ){
-     if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
-     // scale each row (from offset index) in a matrix by a column vector
-     #ifdef _OPENMP
-     #pragma omp parallel for simd shared(src_mtx, src_rows, src_cols, scale_vec, col_start, col_end)
-     #endif
-     for(int i = 0; i< src_rows; i++){  
-          T scale = 1 / scale_vec[i];
-          for(int j=col_start; j<= col_end; j++){
-               src_mtx[src_cols*i + j] = src_mtx[src_cols*i + j] * scale ;
-          }
-     }
-}
+void norm_rows_in_mtx_by_col_vector(T* src_mtx, size_t src_rows, size_t src_cols, T* scale_vec, 
+            long int col_start=0, long int col_end=-1 );
+
 
 
 #if defined (_USE_GSL) || defined (_USE_MKL)
@@ -350,13 +146,7 @@ void norm_rows_in_mtx_by_col_vector<float>(float* src_mtx, size_t src_rows, size
 
 
 template<typename T>
-void norm_rows_by_maxabs_in_each_row(T* src_mtx, size_t src_rows, size_t src_cols, long int max_start_col=0, long int max_end_col=-1, long int norm_start_col =0, long int norm_end_col=-1){     
-     
-     T* norms = new T[src_rows]();     
-     get_max_each_row<T>(norms, src_mtx, src_rows, src_cols, max_start_col, max_end_col);   
-     norm_rows_in_mtx_by_col_vector<T>(src_mtx, src_rows, src_cols, norms, norm_start_col, norm_end_col);          
-     delete[] norms;
-}
+void norm_rows_by_maxabs_in_each_row(T* src_mtx, size_t src_rows, size_t src_cols, long int max_start_col=0, long int max_end_col=-1, long int norm_start_col =0, long int norm_end_col=-1);
